Add key generation function
Adding a generic key generation function `ts_pg_generate_key` for later
use to generate user keys and certificates as well as instance keys and
certificates.  A key checking function `ts_pg_check_key` is also added
which is used to check that the generated key is correct. Both
functions are intended for testing usage and are not part of the public
API.