Prototype for chunk manipulation function
Adding a function `drop_chunk` that can be used to delete an individual
chunk from a hypertable to allow us to experiment with per-chunk
manipulation functions.

Operators `<<` for comparing timestamps with timestamp ranges are
added and the view is changes to provide chunk identifiers and
hypertable identifiers as `regclass` columns and a time range for the
boundaries of the chunk.

To test the function, views `test_info.chunks_tstz` and
`test_info.chunks_ts` are added and tests for `drop_chunk` using these
views are added.