Visibility of job error log lines
A user should only be able to view lines generated by jobs the user
own, or jobs owned by a role that the user is a member of. Since it
should be possible to assign a different owner to a job than the
current user, it is necessary to add a `owner` parameter to `add_job()`
and `alter_job()`.