Make create_hypertable non-experimental
Move experimental functions for creating hypertables and dimensions to
public schema and allow the new and old interface to be overloaded.

This is done by introducing a Dimension object that contain the
information about the dimension and then pass that to the
create_hypertable function. This allow the old and the new interface to
be used interchangably.